// app/api/register/RegistorValidation.ts
export class RegisterValidation {
    InvalidName: boolean = false;
    InvalidEmail: boolean = false;
    InvalidPassword: boolean = false;
    InvalidUserId: boolean = false;
    InvalidPhone: boolean = false;
    InvalidLoc: boolean = false;
    InvalidRegNo: boolean = false;
    error: string = '';
    showPassword: boolean = false;
    showConfirmPass: boolean = false;
    confirmUser: string = '';
    confirmPass: string = '';
  }
  